Dodajem osnovnu logiku 'Utility' projekta
Izracunavanje cene struje u zavisnosti od vremena
Potrebno obezbetiti logiku wcf-a...

+++ RES_SHES_PR-22-27-2015/Common/Helper.cs
@@ -15,7 +15,7 @@ namespace SHES
 
         public static double MinutesToHours(double minutes)
         {
-            return minutes / 60;
+            return minutes / (double)60;
         }
     }
 }
